requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pynput==1.7.6
pyautogui==0.9.54
sounddevice>=0.4.6
numpy>=1.20.0
pystray==0.19.4
pillow==10.1.0
qrcode==7.4.2
pywin32==306
python-multipart==0.0.6
